Move `landing_page` to a new `scanner` app
