+++ landing_page_checker/management/commands/scan.py
@@ -1,7 +1,7 @@
 from django.core.management import BaseCommand, CommandError
 
 from directory.models import DirectoryEntry
-from landing_page_checker.landing_page.scanner import bulk_scan
+from scanner.scanner import bulk_scan
 from landing_page_checker.utils import url_to_domain
