+++ scanner/tests/test_scanner.py
@@ -4,8 +4,8 @@ from unittest import mock
 from django.test import TestCase
 import vcr
 
-from landing_page_checker.landing_page import scanner
-from landing_page_checker.landing_page.tests.utils import (
+from scanner import scanner
+from scanner.tests.utils import (
     NON_EXISTENT_URL,
     requests_get_mock,
 )
@@ -26,7 +26,7 @@ class ScannerTest(TestCase):
     """
 
     @mock.patch(
-        'landing_page_checker.landing_page.scanner.requests.get',
+        'scanner.scanner.requests.get',
         new=requests_get_mock
     )
     @mock.patch(
@@ -141,7 +141,7 @@ class ScannerTest(TestCase):
             )
 
     @mock.patch(
-        'landing_page_checker.landing_page.scanner.requests.get',
+        'scanner.scanner.requests.get',
         new=requests_get_mock
     )
     @mock.patch(
