+++ scanner/tests/utils.py
